Unit conversions (#30)
This allows to specify the conversion of unit of measurement. Currently only time units are supported. If a tuple in a form of {from, to} is passed to :unit option, the measurement function is wrapped and the measurement is converted to the desired unit.

The unit conversion relies on the fact that the measurement is a number, but since it's not checked anywhere, we simply crash with a BadArithmetic when performing the conversion.