Add a summary metric type (#32)
This change adds a statistics metric type. Its purpose is to aggregate measurement values into a set of statistics, e.g. min/max, mean, percentiles etc. It is up to reporters to decide (and document) what statistics are actually expose, because it depends on the monitoring system in use.
Statistics are useful in situations where distributions are not available or we don't know yet what buckets we should choose for the distribution. It also makes sense to use them when there is only one "instance" of the metric (e.g. our application runs only on one node).