Support bytes, kilobytes and megabytes conversion
Before this commit, only time conversions were supported.
Some metrics, like the Erlang VM's memory usage, are reported in bytes.
It might be interesting for some reporters to have the ability to
convert this value a bigger unit.

Note: The International System of Units (SI) definition was used here,
which means that 1 Kilobyte is defined as 1000 bytes instead of 1024.
This definition was chosen to avoid ambiguity, since the 1024 notation
should now be called kibibytes, mebibytes, gibibytes and so on.

References:

- https://physics.nist.gov/cuu/Units/binary.html
- https://en.wikipedia.org/wiki/Kilobyte
- https://cseducators.stackexchange.com/questions/4425/should-i-teach-that-1-kb-1024-bytes-or-1000-bytes