Unified similar tests
This commit follows a similar approach as the one used to test time
conversions, by grouping the same kind of tests and only changing the
input.
It was decided, however, to use static values rather than random
ones like the time conversion tests. The reasoning behind this decision
is that it's easier to reason about each case input and expected output,
and also that there's no need to create helper functions that mimic the
conversion behavior inside the tests.