Use public function instead of lambda
Metrics specifications can be copied to many different processes by some
of the handlers, this mean that we need to reduce the size of metrics
specs as much as possible to perform well in high throughput services.
This change replaces lambda with function capture, that not only
provides smaller footprint, it also allows some handlers to pattern
match on function if needed.